Uploaded the incomplete game
remember if you're making changes consider making your own branch before
uploading to master and on top of that please document well. It is very
hard to figure out what is going on with the current code we have to
work with. We don't have to go and make it any harder than it currently
is.