Added a lot
Added and "Completed" BaseState, Game, StateManager. The program
compiles and does everything I expected it to except for one thing. I
intended for the StateManager class to be a singleton class. Due to me
not being able to make this work I settled on just making it a normal
class. I don't believe this should negatively effect anything. All this
means is that I will have to create an instance of StateManager inside
of the Game class. I may make StateManager a singleton class in the
future but this is not a priority.