A lot again..
I finished the FileManager class. This is a very vital part of the
program and it is completely finished. The SplashScreenState is almost
done. It is really being used for debugging right now. The important
thing is StateManager is also completely done and it works perfectly
with SplashScreenState. At this point I need to flesh out the menu and
Game state classes then from there it should be programming Entities,
staticEnvironment, and dynamicEnvironment AKA easy stuff that doesn't
stress me out. I'm making amazing progress and the end goal seems a lot
less elusive now.